Add required flag to registerReceiver call in RemoteBugreportManager
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
unprotected broadcasts. This commit adds the RECEIVER_EXPORTED flag
to the call to registerReceiver in RemoteBugreportManager when
registering for the REMOTE_BUGREPORT_DISPATCH action as this is
broadcast from shell.

Bug: 161145287
Test: Manually verified logcat warning was not reported
Change-Id: I51ef7b5c275c70edd4eaa98132b98315c1e04256