Apply display transform to window bounds and node bounds
Currently, the node bounds and the window bounds are both in
unrotated coordinates, it means if their layout is not changed,
then the bound wouldn't be change after rotating the device.

This effects Uiobject2, which is used for uiautomator, because
it clips the nodes bounds based on current display metrics.
see ag/13805539.

To fix it, we transfer them back to the rotated coordinates
based on the displayInfo called back from the surface flinger.
(Co-Author: Ryan)

Bug: 191736824
Test: a11y CTS & unit tests
Test: atest NexusLauncherTests & Launcher3Tests
Change-Id: I14c5e98895c370abe11b08c466bb7091f7ea613e