Allow camera to specify a refresh rate range
Instead of using a hard coded 60Hz value, allow the camera service
to specify a range of refresh rate for the display to use. In this CL
we still use hardcoded 60Hz, but we move it to the CameraServiceProxy.
In a follow up CL the camera would provide the real range to use.

Bug: 209669709
Test: atest RefreshRatePolicy
Change-Id: I222dedbdef9f9d0e1f64d01074d2e9a04867b178